+++ ui/apps/platform/src/Containers/NetworkGraph/TopologyComponent.tsx
@@ -33,61 +33,8 @@ import {
 } from './hooks/useNetworkPolicySimulator';
 import { EdgeState } from './components/EdgeStateSelect';
 import { deploymentTabs } from './utils/deploymentUtils';
+import { getInScopeEntities } from './utils/simulatorUtils';
 import { NetworkScopeHierarchy } from './types/networkScopeHierarchy';
-import { EntityScope } from './simulation/NetworkPoliciesGenerationScope';
-
-/**
- * Given the array of nodeModels returned by the network graph API and the user's
- * current scope hierarchy, return the scope of entities that should be included
- * in the network policy simulation.
- */
-function getInScopeEntities(
-    nodeModels: CustomNodeModel[],
-    scopeHierarchy: NetworkScopeHierarchy
-): EntityScope {
-    let granularity: EntityScope['granularity'] = 'CLUSTER';
-
-    if (scopeHierarchy.namespaces.length === 0 && scopeHierarchy.deployments.length === 0) {
-        // Only a cluster is selected
-        granularity = 'CLUSTER';
-    } else if (scopeHierarchy.namespaces.length > 0 && scopeHierarchy.deployments.length === 0) {
-        // A cluster and at least one namespace is selected
-        granularity = 'NAMESPACE';
-    } else {
-        // A cluster, at least one namespace, and at least one deployment is selected
-        granularity = 'DEPLOYMENT';
-    }
-
-    // A cluster, at least one namespace, and at least one deployment is selected. We
-    // include all deployments from the model that match a selected namespace and
-    // selected deployment name from the scope hierarchy.
-    const namespaceNameSet = new Set(scopeHierarchy.namespaces);
-    const deploymentNameSet = new Set(scopeHierarchy.deployments);
-
-    const deploymentScope: EntityScope = {
-        granularity,
-        cluster: scopeHierarchy.cluster.name,
-        namespaces: scopeHierarchy.namespaces,
-        deployments: [],
-    };
-
-    nodeModels.forEach((node) => {
-        if (node.data.type !== 'DEPLOYMENT') {
-            return;
-        }
-
-        const inSelectedNamespaces = namespaceNameSet.has(node.data.deployment.namespace);
-        const inSelectedDeployments = deploymentNameSet.has(node.data.deployment.name);
-        if (
-            (granularity === 'NAMESPACE' && inSelectedNamespaces) ||
-            (granularity === 'DEPLOYMENT' && inSelectedNamespaces && inSelectedDeployments)
-        ) {
-            deploymentScope.deployments.push(node.data.deployment);
-        }
-    });
-
-    return deploymentScope;
-}
 
 // TODO: move these type defs to a central location
 export const UrlDetailType = {
