+++ ui/apps/platform/src/Containers/NetworkGraph/utils/simulatorUtils.ts
@@ -1,4 +1,7 @@
 import { NetworkPolicyModification } from 'types/networkPolicy.proto';
+import { EntityScope } from '../simulation/NetworkPoliciesGenerationScope';
+import { NetworkScopeHierarchy } from '../types/networkScopeHierarchy';
+import { CustomNodeModel } from '../types/topology.type';
 
 export function getDisplayYAMLFromNetworkPolicyModification(
     modification: NetworkPolicyModification | null
@@ -26,4 +29,61 @@ export function getDisplayYAMLFromNetworkPolicyModification(
     }
 
     return displayYaml;
+}
+
+function getGranularityFromScopeHierarchy(
+    scopeHierarchy: NetworkScopeHierarchy
+): EntityScope['granularity'] {
+    if (scopeHierarchy.namespaces.length === 0 && scopeHierarchy.deployments.length === 0) {
+        // Only a cluster is selected
+        return 'CLUSTER';
+    }
+    if (scopeHierarchy.namespaces.length > 0 && scopeHierarchy.deployments.length === 0) {
+        // A cluster and at least one namespace is selected
+        return 'NAMESPACE';
+    }
+    // A cluster, at least one namespace, and at least one deployment is selected
+    return 'DEPLOYMENT';
+}
+
+/**
+ * Given the array of nodeModels returned by the network graph API and the user's
+ * current scope hierarchy, return the scope of entities that should be included
+ * in the network policy simulation.
+ */
+export function getInScopeEntities(
+    nodeModels: CustomNodeModel[],
+    scopeHierarchy: NetworkScopeHierarchy
+): EntityScope {
+    const granularity = getGranularityFromScopeHierarchy(scopeHierarchy);
+
+    // A cluster, at least one namespace, and at least one deployment is selected. We
+    // include all deployments from the model that match a selected namespace and
+    // selected deployment name from the scope hierarchy.
+    const namespaceNameSet = new Set(scopeHierarchy.namespaces);
+    const deploymentNameSet = new Set(scopeHierarchy.deployments);
+
+    const deploymentScope: EntityScope = {
+        granularity,
+        cluster: scopeHierarchy.cluster.name,
+        namespaces: scopeHierarchy.namespaces,
+        deployments: [],
+    };
+
+    nodeModels.forEach((node) => {
+        if (node.data.type !== 'DEPLOYMENT') {
+            return;
+        }
+
+        const inSelectedNamespaces = namespaceNameSet.has(node.data.deployment.namespace);
+        const inSelectedDeployments = deploymentNameSet.has(node.data.deployment.name);
+        if (
+            (granularity === 'NAMESPACE' && inSelectedNamespaces) ||
+            (granularity === 'DEPLOYMENT' && inSelectedNamespaces && inSelectedDeployments)
+        ) {
+            deploymentScope.deployments.push(node.data.deployment);
+        }
+    });
+
+    return deploymentScope;
 }
